+++ metadata.rb
@@ -1,9 +1,11 @@
-name             'users'
+name             'ssh-users'
 maintainer       'YOUR_COMPANY_NAME'
 maintainer_email 'YOUR_EMAIL'
 license          'All rights reserved'
-description      'Installs/Configures users'
+description      'Installs/Configures ssh-users'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+
 depends 'chef-solo-search'
 depends 'users'
+depends 'sudo'